requirements.txt
streamlit
selenium
beautifulsoup4
openai
python-dotenv
langchain
pinecone-client
tqdm
